+++ Userland/SampleCodeModule/include/tron.h
@@ -1,7 +1,9 @@
 #ifndef TRON_H
 #define TRON_H
 #include <stdint.h>
+
 void play_tron();
-void setPlayers(uint64_t width, uint64_t height);
-void setScreen(uint64_t width, uint64_t height);
+void setPlayers(uint16_t width, uint16_t height);
+void setScreen(uint16_t width,uint16_t height);
+
 #endif
